Create component current version with reply to review
